requirements/common.txt
Django==1.5.2
South==0.8.2
django-bootstrap-form==3.0
django-registration==0.8
django-tastypie==0.10.0
mimeparse==0.1.3
python-dateutil==2.1
six==1.4.1
wsgiref==0.1.2
